rename service modules and have them export default
